Mise en place du service de messagerie via starter mail de spring boot

+++ start.sh
@@ -1,5 +1,12 @@
 #!/usr/bin/env bash
+MAIL_RECEIVER=$1
+if [ $# = 0 ]
+then
+    echo "No arguments recieved ! You should specify your email"
+	exit 1 
+fi
 echo "docker-compose down"
+echo "Every email will be sent to $MAIL_RECEIVER"
 docker-compose down --rmi all
 echo "cd api"
 cd api
@@ -10,6 +17,8 @@ cd ../app
 ng build --aot --build-optimizer --prod
 echo "docker-compose up"
 cd ..
-docker-compose up 
+export MAIL_RECEIVER=$MAIL_RECEIVER
+docker-compose up
+# tounga.franck@ng-itconsulting.com
 # force rebuild all image
-# docker-compose up -d --force-recreate --build
+# docker-compose up --force-recreate --build
